+++ src/OFS/FindSupport.py
@@ -68,88 +68,12 @@ class FindSupport(Base):
                  search_sub=0,
                  REQUEST=None, result=None, pre=''):
         """Zope Find interface"""
-
-        if result is None:
-            result=[]
-
-            if obj_metatypes and 'all' in obj_metatypes:
-                obj_metatypes=None
-
-            if obj_mtime and type(obj_mtime)==type('s'):
-                obj_mtime=DateTime(obj_mtime).timeTime()
-
-            if obj_permission:
-                obj_permission=p_name(obj_permission)
-
-            if obj_roles and type(obj_roles) is type('s'):
-                obj_roles=[obj_roles]
-
-            if obj_expr:
-                # Setup expr machinations
-                md=td()
-                obj_expr=(Eval(obj_expr), md, md._push, md._pop)
-
-        base = aq_base(obj)
-
-        if hasattr(base, 'objectItems'):
-            try:    items=obj.objectItems()
-            except: return result
-        else:
-            return result
-
-        try: add_result=result.append
-        except:
-            raise AttributeError, `result`
-
-        for id, ob in items:
-            if pre: p="%s/%s" % (pre, id)
-            else:   p=id
-
-            dflag=0
-            if hasattr(ob, '_p_changed') and (ob._p_changed == None):
-                dflag=1
-
-            bs = aq_base(ob)
-            if (
-                (not obj_ids or absattr(bs.getId()) in obj_ids)
-                and
-                (not obj_metatypes or (hasattr(bs, 'meta_type') and
-                 bs.meta_type in obj_metatypes))
-                and
-                (not obj_searchterm or
-                 (hasattr(ob, 'PrincipiaSearchSource') and
-                  ob.PrincipiaSearchSource().find(str(obj_searchterm)) >= 0
-                  )
-                 or
-                 (hasattr(ob, 'SearchableText') and
-                  ob.SearchableText().find(str(obj_searchterm)) >= 0)
-                 )
-                and
-                (not obj_expr or expr_match(ob, obj_expr))
-                and
-                (not obj_mtime or mtime_match(ob, obj_mtime, obj_mspec))
-                and
-                ( (not obj_permission or not obj_roles) or \
-                   role_match(ob, obj_permission, obj_roles)
-                )
-                ):
-                add_result((p, ob))
-                dflag=0
-
-            if search_sub and (hasattr(bs, 'objectItems')):
-                subob = ob
-                sub_p = p
-                self.ZopeFind(subob, obj_ids, obj_metatypes,
-                                   obj_searchterm, obj_expr,
-                                   obj_mtime, obj_mspec,
-                                   obj_permission, obj_roles,
-                                   search_sub,
-                                   REQUEST, result, sub_p)
-            if dflag: ob._p_deactivate()
-
-        return result
-
-
+        return self.ZopeFindAndApply(obj, obj_ids=obj_ids,
+                  obj_metatypes=obj_metatypes, obj_searchterm=obj_searchterm,
+                  obj_expr=obj_expr, obj_mtime=obj_mtime, obj_mspec=obj_mspec,
+                  obj_permission=obj_permission, obj_roles=obj_roles,
+                  search_sub=search_sub, REQUEST=REQUEST, result=result,
+                  pre=pre, apply_func=None, apply_path='')
 
     security.declareProtected(view_management_screens, 'PrincipiaFind')
     PrincipiaFind=ZopeFind
@@ -212,8 +136,11 @@ class FindSupport(Base):
                 and
                 (not obj_searchterm or
                  (hasattr(ob, 'PrincipiaSearchSource') and
-                  ob.PrincipiaSearchSource().find(obj_searchterm) >= 0
-                  ))
+                  obj_searchterm in ob.PrincipiaSearchSource())
+                 or 
+                 (hasattr(ob, 'SearchableText') and
+                  obj_searchterm in ob.SearchableText())
+                 )
                 and
                 (not obj_expr or expr_match(ob, obj_expr))
                 and